Add files via upload
Processing the embedded patches and learning the complex features and patterns within the image. nn.TransformerEncderLayer = represents a single layer of a Transformer Encoder. The transformer encoder consist of alternating layers of multi head self attention and MLP blocks, layernorm(LN) is applied before every block and residual connection after every block. MLP(Multi-layer perceptron=Feed forward layer) contains two layers with a GELU and Non-linearity. Work on image from top to bottom, starting with the MLP. MLP is said to contain two layers with a GELU non-linearity. To construct the MLP , we want the second layer to have larger dimension than the first one. By hyper parameters, we set an EXPANSION=128, means first layer will have an input dimension of 16 and an output dimension of 16*128=2048. Second layer then have input dimension of 2048 and and output dimension of 16. After each layer we will apply GELU activation and a dropout.